# [Performance] Use new AWS custom images