<!-- Describe the changes in this PR -->
Use the new custom image for AWS's VM creation. 

Benchmark results:
|  VM Type 💻   | Old Provision 🕐                 | New Provision 🕐     | % Improvement ✅ |
| --------------- | ---------------------------- | ----------- | ----------- |
| AWS GPU      | 2min 27s      | 1min 41s    | 31% |
| AWS CPU       | 2min 44s   | 1min 13s | 56% |
* Provision Time = from start of launching an instance to instance setup and ready to run any user workloads (does not include time to search for the available instance)


<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [x]  Run all smoke tests `pytest tests/test_smoke.py --aws`