@yika-luo 2024-10-16
This PR depends on https://github.com/skypilot-org/skypilot-catalog/pull/89